+++ app/layout.js
@@ -1,6 +1,7 @@
 import { Inter } from "next/font/google";
 import "./globals.css";
 import { ThemeProvider } from '@mui/material/styles';
+import { ClerkProvider } from "@clerk/nextjs";
 import theme from '../theme.js'
 
 const inter = Inter({ subsets: ["latin"] });
@@ -12,12 +13,14 @@ export const metadata = {
 
 export default function RootLayout({ children }) {
   return (
+    <ClerkProvider>
     <html lang="en">
       <body className={inter.className}>
-        <ThemeProvider theme={theme}>
-            {children}
-        </ThemeProvider>
-        </body>
+          <ThemeProvider theme={theme}>
+              {children}
+          </ThemeProvider>
+      </body>
     </html>
+  </ClerkProvider>
   );
 }
