+++ package.json
@@ -9,15 +9,17 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@clerk/nextjs": "^5.3.3",
     "@emotion/react": "^11.13.3",
     "@emotion/styled": "^11.13.0",
     "@mui/icons-material": "^5.16.7",
     "@mui/material": "^5.16.7",
-    "@pinecone-database/pinecone": "^3.0.0",
+    "@pinecone-database/pinecone": "^3.0.1",
     "axios": "^1.7.4",
     "cheerio": "^1.0.0",
     "next": "14.2.5",
     "openai": "^4.56.0",
+    "pinecone": "^0.1.0",
     "react": "^18",
     "react-dom": "^18"
   },
